refactor: SOLID splits — separate concerns in large files
- aggregate/redemeineComponent.ts (257→131 lines): extract commandFactory.ts
  (77 lines) and componentTypes.ts (92 lines)
- mirage/createDemeineBridge.ts (252→196 lines): extract bridgeAdapter.ts
  (72 lines) for string utilities and method map builders

Each file now has a single clear responsibility. Backward-compatible
re-exports maintain existing import paths.

Part of: redemeine-o3g (Phase 2 SOLID)
