+++ packages/aggregate/src/redemeineComponent.ts
@@ -1,169 +1,43 @@
-import { type CommandContext, type CommandIntents, type CommandResult, type Event, type PluginExtensions, type SelectorsMap, type ReadonlyDeep } from '@redemeine/kernel';
-import { createCommandContextProxy } from './proxies/createCommandContextProxy';
-import type { Merge } from './types/Merge';
-import type { AllKeys } from './types/AllKeys';
-import type { ReplaceFirstArg } from './types/ReplaceFirstArg';
-import type { PackedCommandWithMeta, ShorthandCommandWithMeta } from './types/aggregateTyping';
+import type { SelectorsMap } from '@redemeine/kernel';
+import type {
+  RedemeineEventDefinition,
+  AnyFunction,
+  ComponentBehaviorSnapshot,
+  InheritableComponentBehavior
+} from './componentTypes';
+import { composeCommandFactories, type GenericCommandFactory } from './commandFactory';
 
-export type GenericSelectors = Record<string, unknown>;
-// SAFETY: `any` required — GenericCommandMap must accept commands with any state type (contravariant)
-export type GenericCommandMap = Record<string, RedemeineCommandDefinition<any, Record<string, unknown>, {}>>;
-export type GenericCommandFactoryContext<TCommands extends Record<string, unknown> = Record<string, unknown>> = {
-  selectors: GenericSelectors;
-  commands?: CommandContext<CommandIntents<TCommands>>;
-  plugins?: Record<string, unknown>;
-};
-export type GenericCommandFactory = (emit: unknown, context: GenericCommandFactoryContext) => GenericCommandMap;
+// Re-export everything from sub-modules for backward compatibility
+export type {
+  GenericSelectors,
+  GenericCommandMap,
+  GenericCommandFactoryContext,
+  GenericCommandFactory,
+  RedemeineShorthandCommand,
+  RedemeineCommandDefinition,
+  RedemeineCommandMap
+} from './commandFactory';
+export {
+  resolveCommandFactoryContext,
+  composeCommandFactories,
+  resolveCommandHandler,
+  createCommandPayload
+} from './commandFactory';
 
-export function resolveCommandFactoryContext(
-  context: GenericCommandFactoryContext
-): GenericCommandFactoryContext {
-  if (context.commands) {
-    return context;
-  }
-
-  return {
-    ...context,
-    commands: createCommandContextProxy<Record<string, unknown>>()
-  };
-}
-
-// SAFETY: `any` in Event type params required — projectors accept events with heterogeneous payloads
-export type RedemeineEventProjector<S> = (state: S, event: Event<any, any>) => void;
-export type RedemeineEventDefinition<S, TMeta extends Record<string, unknown> = Record<string, unknown>> =
-  | RedemeineEventProjector<S>
-  | {
-      projector: RedemeineEventProjector<S>;
-      meta?: TMeta;
-    };
-
-export type NormalizeEventDefinitions<T extends Record<string, RedemeineEventDefinition<any, any>>> = {
-  // SAFETY: `any` in conditional extends clauses required for inference of arbitrary function shapes
-  [K in keyof T]: T[K] extends (...args: any[]) => any
-    ? T[K]
-    : T[K] extends { projector: infer P }
-      ? Extract<P, (...args: any[]) => any>
-      : never;
-};
-
-export type RedemeineShorthandCommand<S, Args extends unknown[] = unknown[], TPlugins extends PluginExtensions = {}> = (
-  state: ReadonlyDeep<S>,
-  ...args: Args
-) => Event<unknown, string> | CommandResult<Event<unknown, string>, TPlugins>;
-
-export type RedemeineCommandDefinition<
-  S,
-  TMeta extends Record<string, unknown> = Record<string, unknown>,
-  TPlugins extends PluginExtensions = {}
-> =
-  // SAFETY: `any[]` in these positions required for contravariant argument matching in command definitions
-  | RedemeineShorthandCommand<S, any[], TPlugins>
-  | ShorthandCommandWithMeta<S, any[], TMeta, TPlugins>
-  | PackedCommandWithMeta<S, any[], unknown, TMeta, TPlugins>;
-
-export type RedemeineCommandMap<S, TMeta extends Record<string, unknown> = Record<string, unknown>, TPlugins extends PluginExtensions = {}> = Record<string, RedemeineCommandDefinition<S, TMeta, TPlugins>>;
-
-export interface RedemeineComponent<
-  S,
-  // SAFETY: `any` in generic defaults required for structural subtyping of component unions
-  Commands extends Record<string, any> = {},
-  Events extends Record<string, any> = {},
-  Projectors extends Record<string, any> = {},
-  Selectors extends Record<string, any> = {},
-  EventOverrides extends object = Record<string, string>,
-  CommandOverrides extends object = Record<string, string>
-> {
-  /**
-   * Phantom generic brand to enforce component state compatibility in structural typing.
-   */
-  readonly __stateType?: S;
-  readonly state?: ReadonlyDeep<unknown>;
-  readonly commands: Commands;
-  readonly events: Events;
-  readonly projectors: Projectors;
-  readonly selectors: Selectors;
-  readonly eventOverrides: EventOverrides;
-  readonly commandOverrides: CommandOverrides;
-}
-
-// SAFETY: `any` in component type positions required for covariant union extraction
-export type ComponentCommandUnion<T extends readonly RedemeineComponent<any, any, any, any, any>[]> =
-  T[number] extends RedemeineComponent<any, infer C, any, any, any> ? C : {};
-
-export type MergeComponentCommands<T extends readonly RedemeineComponent<any, any, any, any, any>[]> =
-  Merge<ComponentCommandUnion<T>>;
-
-export type MergeComponentCommandKeys<T extends readonly RedemeineComponent<any, any, any, any, any>[]> =
-  AllKeys<ComponentCommandUnion<T>>;
-
-export type PublicCommandArgsFromDefinition<S, TDef> =
-  TDef extends { pack: (...args: infer A) => unknown }
-    ? A
-    : ReplaceFirstArg<never, Extract<TDef, (state: ReadonlyDeep<S>, ...args: unknown[]) => unknown>> extends (
-        state: never,
-        ...args: infer A
-      ) => unknown
-      ? A
-      : never;
-
-export type PublicCommandMethodsFromInternal<S, TCommands extends Record<string, unknown>, TResult> = {
-  [K in keyof TCommands]: (...args: PublicCommandArgsFromDefinition<S, TCommands[K]>) => TResult;
-};
-
-export function composeCommandFactories(
-  factories: GenericCommandFactory[]
-): GenericCommandFactory {
-  return (emit: unknown, context: GenericCommandFactoryContext) => {
-    const resolvedContext = resolveCommandFactoryContext(context);
-    const merged: GenericCommandMap = {};
-    for (const factory of factories) {
-      Object.assign(merged, factory(emit, resolvedContext));
-    }
-    return merged;
-  };
-}
-
-export function resolveCommandHandler<S>(
-  commandDef: RedemeineCommandDefinition<S>
-): (state: ReadonlyDeep<S>, payload: unknown) => Event<unknown, string> | CommandResult<Event<unknown, string>, {}> {
-  const handler = typeof commandDef === 'function'
-    ? commandDef
-    : commandDef.handler;
-
-  return handler as (
-    state: ReadonlyDeep<S>,
-    payload: unknown
-  ) => Event<unknown, string> | CommandResult<Event<unknown, string>, {}>;
-}
-
-export function createCommandPayload<S>(commandDef: RedemeineCommandDefinition<S>, args: unknown[]): unknown {
-  if (typeof commandDef !== 'function' && 'pack' in commandDef && typeof commandDef.pack === 'function') {
-    return commandDef.pack(...args);
-  }
-  return args[0];
-}
-
-// SAFETY: Using `Function` for event/selector storage because projectors have heterogeneous signatures
-// that are incompatible with a single typed function signature due to contravariance.
-// eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
-export type AnyFunction = Function;
-
-export interface ComponentBehaviorSnapshot<S> {
-  events: Record<string, AnyFunction>;
-  eventMetadata: Record<string, Record<string, unknown> | undefined>;
-  eventOverrides: Record<string, string>;
-  selectors: SelectorsMap<S>;
-  commandOverrides: Record<string, string>;
-}
-
-export interface InheritableComponentBehavior {
-  events: Record<string, AnyFunction>;
-  eventMetadata: Record<string, Record<string, unknown> | undefined>;
-  eventOverrides: Record<string, string>;
-  selectors: Record<string, AnyFunction>;
-  commandOverrides: Record<string, string>;
-  commandsFactory: GenericCommandFactory;
-}
+export type {
+  RedemeineEventProjector,
+  RedemeineEventDefinition,
+  NormalizeEventDefinitions,
+  RedemeineComponent,
+  ComponentCommandUnion,
+  MergeComponentCommands,
+  MergeComponentCommandKeys,
+  PublicCommandArgsFromDefinition,
+  PublicCommandMethodsFromInternal,
+  AnyFunction,
+  ComponentBehaviorSnapshot,
+  InheritableComponentBehavior
+} from './componentTypes';
 
 export function createComponentBehaviorState<S>() {
   let events: Record<string, AnyFunction> = {};
